requirements.txt
requests
beautifulsoup4
glob2
bytesbufio
pillow

keras
numpy
opencv-python


